Moved home page logic to Home.js and fixed routing logic
