--- frontend/src/components/NavBar/index.js
@@ -1,1 +1,0 @@
-export { default } from "./NavBar";
